api/pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = ""
authors = ["Kim Minh Thang <thangved02@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.108.0"
uvicorn = {extras = ["standard"], version = "^0.25.0"}
python-multipart = "^0.0.6"
transformers = {extras = ["torch"], version = "^4.36.2"}
pytesseract = "^0.3.10"
pdf2image = "^1.17.0"
openai = "^1.6.1"
python-dotenv = "^1.0.0"
langchain = "^0.1.0"
langchain-openai = "^0.0.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
